# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/cards": {
            "put": {
                "description": "Метод API, позволяющий авторизированному пользователю изменить информацию об электронном ключе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Обновление информации электронного ключа",
                "parameters": [
                    {
                        "description": "Тело запроса формата 'application/json', содержащее информацию для обновления ключа",
                        "name": "UpdateKeyBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/integrserv.KeyData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа запроса обновления ключа",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.KeyData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Структура неудачного ответа запроса обновления ключа",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Метод API, позволяющий авторизированному пользователю зарегестрировать новый электронный ключ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Регистрация электронного ключа",
                "parameters": [
                    {
                        "description": "Тело запроса формата 'application/json', содержащее информацию о добавляемом ключе",
                        "name": "NewKeyBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/integrserv.KeyData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа запроса добавления нового ключа",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.KeyData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Структура неудачного ответа запроса добавления нового ключа",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cards/": {
            "get": {
                "description": "Метод API, позволяющий авторизированному пользователю получить список ключей начиная с шага смещения, указанного в параметре 'offset', количества, заданного параметром 'count'",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Получение ключей СКУД",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Шаг смещения",
                        "name": "offset",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Количество",
                        "name": "count",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа запроса получения ключей",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/integrserv.KeyData"
                                            }
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Структура неудачного ответа запроса получения ключей",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cards/by_card_number": {
            "get": {
                "description": "Метод API, позволяющий авторизированному пользователю получить информацию о ключе по переданному параметру электронного кода 'card_no'",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Получение данных о ключе СКУД",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"CA00000082942101\"",
                        "description": "Электронный код пропуска",
                        "name": "card_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа запроса получения данных о ключе",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.KeyData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Структура неудачного ответа запроса получения данных о ключе",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cards/read_card_number": {
            "get": {
                "description": "Метод API, позволяющий авторизированному пользователю считать электронный код пропуска со считывателя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Считывание электронного кода",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 2,
                        "description": "Номер считывателя",
                        "name": "id_reader",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа запроса получения электронного кода пропуска",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.KeyData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Структура неудачного ответа запроса получения электронного кода пропуска",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cards/wiegand_to_touch_memory": {
            "post": {
                "description": "Метод API, позволяющий авторизированному пользователю конвертировать десятичный код на личецвой стороне электронного пропуска к формату, понятному 'Орион Про' - Wiegand 26",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Конвертация десятичного кода карточки к формату Wiegand 26",
                "parameters": [
                    {
                        "description": "Тело запроса формата 'application/json', содержащее информацию для конвертации электронного кода",
                        "name": "ConvertKeyCodeToWiegandBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqs.WiegandToTouchMemory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа запроса конвертации кода ключа",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Структура неудачного ответа запроса конвертации кода ключа",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "integrserv.KeyData": {
            "type": "object",
            "properties": {
                "accessLevelId": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "codeType": {
                    "type": "integer"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isBlocked": {
                    "type": "boolean"
                },
                "isInStopList": {
                    "type": "boolean"
                },
                "isStoreInDevice": {
                    "type": "boolean"
                },
                "isStoreInS2000": {
                    "type": "boolean"
                },
                "personId": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
...
        "reqs.WiegandToTouchMemory": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "codeSize": {
                    "type": "integer"
                }
            }
        },
...
        "response.Body": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {}
            }
        }
    }
}